test: relocate coverage-chasing tests; add mutation job + suite-hygiene gate (#206)
The `test_coverage_100.py` ("Tests to achieve 100% coverage") and
`test_coverage_gaps.py` buckets were coverage-chasing rather than behavior-first.
Relocate every test into the module it exercises:
- console.py  -> tests/test_console.py (new)
- _shared.py  -> tests/test_shared.py (new)
- cli.py      -> tests/test_cli_commands.py
- bump.py     -> tests/test_bump_command.py
- release.py  -> tests/test_release_command.py (+ the former gaps tests)
- rollback.py -> tests/test_rollback_command.py

Add tests/test_suite_hygiene.py: forbids `test_coverage*.py` buckets and
"achieve … coverage" purpose docstrings, so the smell cannot return.

Wire mutation testing: add mutmut (2.x, matching the `make mutation` flags) to
the test group and a scheduled/manual `.github/workflows/mutation.yml`. It is
intentionally non-blocking — it establishes the mutation-score baseline; a
`--fail-under` threshold will be promoted into CI once that baseline is known.

Coverage stays 100% (436 passed); no behavior change.

Closes #206
