+++ .github/workflows/mutation.yml
@@ -1,0 +1,50 @@
+name: "Mutation testing (mutmut)"
+
+# Mutation testing re-runs the test suite once per mutant, so it is far too slow
+# to be a required per-PR gate. It runs weekly and on demand to establish and
+# track the mutation score (how many injected faults the tests actually catch).
+# It is intentionally non-blocking: once a baseline score is known, a
+# `--fail-under`-style threshold can be promoted into the required CI. See #206.
+
+on:
+  schedule:
+    - cron: "0 5 * * 1" # Mondays 05:00 UTC
+  workflow_dispatch:
+
+permissions:
+  contents: read
+
+concurrency:
+  group: mutation-${{ github.ref }}
+  cancel-in-progress: true
+
+jobs:
+  mutation:
+    name: Run mutation tests
+    runs-on: ubuntu-latest
+    timeout-minutes: 60
+    steps:
+      - name: Harden the runner (Audit all outbound calls)
+        uses: step-security/harden-runner@9af89fc71515a100421586dfdb3dc9c984fbf411 # v2.19.4
+        with:
+          egress-policy: audit
+
+      - name: Checkout Code
+        uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
+
+      - name: Install uv
+        uses: astral-sh/setup-uv@37802adc94f370d6bfd71619e3f0bf239e1f3b78 # v7.6.0
+
+      - name: Run mutation tests (informational, non-blocking)
+        # Surviving mutants must not fail this job; it exists to report the score,
+        # not to gate. A threshold will be enforced once a baseline is established.
+        continue-on-error: true
+        run: make mutation
+
+      - name: Upload mutation HTML report
+        if: always()
+        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
+        with:
+          name: mutation-report
+          path: _tests/mutation/html
+          if-no-files-found: ignore
